requirements.txt
fastapi
uvicorn[standard]
python-dotenv

streamlit
click<=8.0.4
httpx-oauth
ecs-logging
elasticsearch>=8.2.0
sentence_transformers
git+https://github.com/hunkim/streamlit-google-oauth
python-multipart